+++ history.md
@@ -6,6 +6,7 @@
 * Tweaked default options for fugassem_generate_annotation_input utility
 * Update demo run
 * Tweaked workflow name
+* Updated prepare_annotation module
 * Release v0.3.5
 
 ## v0.3.4 2022-11-11 ##