+++ history.md
@@ -1,6 +1,9 @@
 
 # FUGAsseM History #
 
+## v0.3.6 2024-03-25 ##
+* Added option for hyperparameter tuning RF
+
 ## v0.3.5 2023-10-17 ##
 * Fixed issue when using '--bypass-coexp' option
 * Tweaked default options for fugassem_generate_annotation_input utility